reconstract the menu items, make some submenu; add balance; add more output messages; add info methods

+++ ExpensesAndIncome/Expenses.cs
@@ -5,11 +5,11 @@ namespace ExpensesAndIncome;
 public class Expense : IExpensesOrIncome
 {
     public DateTime DataOfAction { get; private set; }
-    public int Amount { get; private set; }
+    public double Amount { get; private set; }
     public string StringTypeOfExpenses { get; private set; }
     public string Comment { get; private set; }
 
-    public Expense(DateTime dataOfAction, int amount, string stringTypeOfExpenses, string comment)
+    public Expense(DateTime dataOfAction, double amount, string stringTypeOfExpenses, string comment)
     {
         DataOfAction = dataOfAction;
         Amount = amount;
