+++ ExpensesAndIncome/IExpensesOrIncome.cs
@@ -5,7 +5,7 @@ namespace ExpensesAndIncome;
 public interface IExpensesOrIncome
 {
     public DateTime DataOfAction { get;  }
-    public int Amount { get;  }
+    public double Amount { get;  }
     public string StringTypeOfExpenses { get; }
     public string Comment { get;  }
