+++ ExpensesAndIncome/TypeOfExpenses.cs
@@ -6,7 +6,7 @@ public class TypeOfExpenses : ITypeOfExpensesAndIncome
 {
 
     public string NameOfType { get;  set; }
-    public int TotalSummOfType { get;  set; }
+    public double TotalSummOfType { get;  set; }
 
     public TypeOfExpenses(string nameOfType)
     {
