+++ ExpensesAndIncome/TypeOfIncome.cs
@@ -5,7 +5,7 @@ namespace ExpensesAndIncome;
 public class TypeOfIncomes : ITypeOfExpensesAndIncome
 {
     public string NameOfType { get;  set; }
-    public int TotalSummOfType { get;  set; }
+    public double TotalSummOfType { get;  set; }
     
 
     public TypeOfIncomes(string nameOfType)
